# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Xolo Ping API
  description: |
    A minimal spec-first ping/pong API with WorkOS AuthKit authentication. This
    file is the single source of truth for both the Go server stub (generated by
    ogen) and the TypeScript client (generated by openapi-typescript).

    Note: the browser-redirect auth routes (`GET /auth/login`, `GET /auth/callback`,
    `GET /auth/logout`) are intentionally NOT modelled here. They are 302 redirect
    flows that set/clear an HttpOnly session cookie, not JSON operations, so they
    live as plain net/http handlers alongside the generated server. Only the JSON
    API (the operations below) is spec-driven.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /ping:
    get:
      operationId: ping
      summary: Ping the server
      description: |
        Returns a pong message. Requires an authenticated session — the request
        must carry a valid `wos_session` cookie. Used to verify the end-to-end
        authenticated transport.
      responses:
        "200":
          description: A pong response.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PongResponse"
        "401":
          description: No valid session — the caller is not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/verify-email:
    post:
      operationId: verifyEmail
      summary: Complete email-verification during login
      description: |
        Some providers (notably GitHub OAuth) return an unverified email on first
        login, so WorkOS requires email verification before issuing a session. The
        `/auth/callback` handler detects this, stores the WorkOS
        `pending_authentication_token` in a short-lived sealed cookie, and sends
        the browser to the SPA's verification step. The SPA collects the code
        WorkOS emailed and POSTs it here to finish authentication and establish
        the session.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyEmailRequest"
      responses:
        "200":
          description: Verification succeeded; the session cookie is now set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: |
            No pending verification, an expired/invalid pending token, or a wrong
            verification code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /me:
    get:
      operationId: getMe
      summary: Get the current user
      description: |
        Returns the WorkOS user backing the current session. Requires a valid
        `wos_session` cookie; returns 401 when unauthenticated. The SPA calls this
        on load to decide whether to show the signed-in or signed-out UI.
      responses:
        "200":
          description: The authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: No valid session — the caller is not authenticated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/pending-orgs:
    get:
      operationId: getPendingOrgs
      summary: List organizations awaiting selection
      description: |
        During the org-selection step the SPA calls this to read the organizations
        the user may choose between. The choices were stashed by `/auth/callback`
        in a short-lived sealed cookie. Returns 401 if there is no pending
        selection.
      responses:
        "200":
          description: The organizations to choose from.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PendingOrganizations"
        "401":
          description: No pending organization selection.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/select-org:
    post:
      operationId: selectOrg
      summary: Complete login by choosing an organization
      description: |
        Finishes a login that WorkOS gated on organization selection. Exchanges
        the chosen organization plus the stashed pending token for a session,
        sets the session cookie, and returns the user.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SelectOrgRequest"
      responses:
        "200":
          description: Selection succeeded; the session cookie is now set.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          description: No pending selection, or an invalid/expired pending token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /invitations:
    get:
      operationId: listInvitations
      summary: List the active organization's invitations
      description: |
        Returns the invitations for the caller's active organization. Requires a
        session scoped to an organization.
      responses:
        "200":
          description: The organization's invitations.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationListResponse"
        "401":
          description: No valid session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: createInvitation
      summary: Invite an email to the active organization
      description: |
        Sends a WorkOS invitation for the given email to the caller's active
        organization (optionally with a role). WorkOS emails the invitee a link;
        accepting it (by logging in with the invitation token) creates their
        membership.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvitationRequest"
      responses:
        "200":
          description: The created invitation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationResponse"
        "400":
          description: The request was invalid (e.g. no active organization).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: No valid session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /integrations/status:
    get:
      operationId: getIntegrationStatus
      summary: Integration connection status for the active organization
      description: |
        Returns the connection state of each supported integration (GitHub,
        Slack) for the caller's active organization. Drives both the onboarding
        wizard and the integrations settings view. The actual connect/callback
        flows are browser redirects (GET /integrations/{provider}/connect) and
        are not part of this JSON spec.
      responses:
        "200":
          description: Per-provider connection status.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationStatusResponse"
        "401":
          description: No valid session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /integrations/{provider}/disconnect:
    post:
      operationId: disconnectIntegration
      summary: Disconnect an integration from the active organization
      description: Removes the stored installation/token for the given provider.
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
            enum: [github, slack]
      responses:
        "200":
          description: Disconnected (or was already not connected).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IntegrationStatusResponse"
        "401":
          description: No valid session.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
components:
  schemas:
    PongResponse:
      type: object
      description: The response returned by the ping endpoint.
      properties:
        message:
          type: string
          description: Always the literal string "pong".
          example: pong
      required:
        - message
    UserResponse:
      type: object
      description: The authenticated WorkOS user and their active organization context.
      properties:
        id:
          type: string
          description: The WorkOS user ID.
          example: user_01H...
        email:
          type: string
          description: The user's email address.
          example: jane@example.com
        firstName:
          type: string
          description: The user's first name, if set.
          example: Jane
        lastName:
          type: string
          description: The user's last name, if set.
          example: Doe
        organizationId:
          type: string
          description: The ID of the organization the current session is scoped to, if any.
          example: org_01H...
        role:
          type: string
          description: The user's role within the active organization, if any.
          example: admin
        organizations:
          type: array
          description: All organizations the user is a member of.
          items:
            $ref: "#/components/schemas/Organization"
      required:
        - id
        - email
    Organization:
      type: object
      description: A WorkOS organization the user belongs to.
      properties:
        id:
          type: string
          description: The WorkOS organization ID.
          example: org_01H...
        name:
          type: string
          description: The organization's display name.
          example: Acme Inc
        role:
          type: string
          description: The user's role in this organization, if known.
          example: member
      required:
        - id
        - name
    VerifyEmailRequest:
      type: object
      description: The verification code the user received by email.
      properties:
        code:
          type: string
          description: The one-time verification code WorkOS emailed to the user.
          example: "123456"
      required:
        - code
    SelectOrgRequest:
      type: object
      description: The organization the user chose during the org-selection step.
      properties:
        organizationId:
          type: string
          description: The ID of the organization to authenticate into.
          example: org_01H...
      required:
        - organizationId
    PendingOrganizations:
      type: object
      description: The organizations a user may choose between during org selection.
      properties:
        organizations:
          type: array
          items:
            $ref: "#/components/schemas/Organization"
      required:
        - organizations
    CreateInvitationRequest:
      type: object
      description: Invite an email address to the caller's active organization.
      properties:
        email:
          type: string
          description: The email address to invite.
          example: newteammate@example.com
        role:
          type: string
          description: Optional role slug to grant the invitee on acceptance.
          example: member
      required:
        - email
    InvitationResponse:
      type: object
      description: A WorkOS organization invitation.
      properties:
        id:
          type: string
          example: invitation_01H...
        email:
          type: string
          example: newteammate@example.com
        state:
          type: string
          description: The invitation state (pending, accepted, expired, revoked).
          example: pending
        expiresAt:
          type: string
          description: ISO 8601 expiry timestamp.
          example: "2026-07-03T00:00:00Z"
      required:
        - id
        - email
        - state
    InvitationListResponse:
      type: object
      description: A list of invitations for the active organization.
      properties:
        invitations:
          type: array
          items:
            $ref: "#/components/schemas/InvitationResponse"
      required:
        - invitations
    IntegrationStatus:
      type: object
      description: The connection state of a single integration provider.
      properties:
        provider:
          type: string
          description: The provider key.
          example: github
        connected:
          type: boolean
          description: Whether this provider is connected for the active organization.
        account:
          type: string
          description: The connected account/workspace label (GitHub login, Slack team).
          example: acme
        connectedBy:
          type: string
          description: The WorkOS user id who connected it, if known.
          example: user_01H...
      required:
        - provider
        - connected
    IntegrationStatusResponse:
      type: object
      description: Connection status for every supported integration.
      properties:
        integrations:
          type: array
          items:
            $ref: "#/components/schemas/IntegrationStatus"
        configured:
          type: boolean
          description: Whether the integration service is configured (database available) on the server.
      required:
        - integrations
        - configured
    Error:
      type: object
      description: A generic error response.
      properties:
        message:
          type: string
          description: A human-readable description of the error.
          example: unauthorized
      required:
        - message